+++ CHANGELOG.md
@@ -5,11 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.18.1.3] - 2025-05-21
+## [1.18.1.4] - 2025-05-26
 
 ### Fixed
 
-- Fix bug with page/tag title not being untranslated
+- Fix bug with audio (dubbing) not always untranlating on the first load (due to persistency of player.unstranslatedAudio)
+
+### Changes
+
+- Do not process elements outside the viewport or not visible to avoid slowing the page
+- Use IntersectObserver to dynamically work on element entering the viewport
+
+## [1.18.1.3] - 2025-05-23
+
+### Fixed
+
+- Fix bug with page/tab title not being untranslated
 
 ## [1.18.1.2] - 2025-05-20